Fix HttpClient using expired client certificate when certificate rotation happens
